Restructure solution.

+++ Advent of Code 2024/14/space.cpp
@@ -25,20 +25,20 @@ size_t Space::calculateSafetyFactorAfter(size_t seconds)
 	size_t	quad3Num	{ 0 };
 	size_t	quad4Num	{ 0 };
 
-	for ( int i = 0, height = fSize.fY / 2; i < height; i++ )
-		for ( int j = fSize.fX / 2 + 1; j < fSize.fX; j++ )
+	for ( int64_t i = 0, height = fSize.fY / 2; i < height; i++ )
+		for ( int64_t j = fSize.fX / 2 + 1; j < fSize.fX; j++ )
 			quad1Num	+= fRobotGrid[ i ][ j ];
 
-	for ( int i = 0, height = fSize.fY / 2; i < height; i++ )
-		for ( int j = 0, width = fSize.fX / 2; j < width; j++ )
+	for ( int64_t i = 0, height = fSize.fY / 2; i < height; i++ )
+		for ( int64_t j = 0, width = fSize.fX / 2; j < width; j++ )
 			quad2Num	+= fRobotGrid[ i ][ j ];
 
-	for ( int i = fSize.fY / 2 + 1; i < fSize.fY; i++ )
-		for ( int j = 0, width = fSize.fX / 2; j < width; j++ )
+	for ( int64_t i = fSize.fY / 2 + 1; i < fSize.fY; i++ )
+		for ( int64_t j = 0, width = fSize.fX / 2; j < width; j++ )
 			quad3Num	+= fRobotGrid[ i ][ j ];
 
-	for ( int i = fSize.fY / 2 + 1; i < fSize.fY; i++ )
-		for ( int j = fSize.fX / 2 + 1; j < fSize.fX; j++ )
+	for ( int64_t i = fSize.fY / 2 + 1; i < fSize.fY; i++ )
+		for ( int64_t j = fSize.fX / 2 + 1; j < fSize.fX; j++ )
 			quad4Num	+= fRobotGrid[ i ][ j ];
 
 	return quad1Num * quad2Num * quad3Num * quad4Num;
@@ -47,11 +47,11 @@ size_t Space::calculateSafetyFactorAfter(size_t seconds)
 // inspired by reddit, didn't pan out.
 size_t Space::findLowestSafetyFactorConfiguration()
 {
-	const int			iterations		{ fSize.fX * fSize.fY };
+	const int64_t		iterations		{ fSize.fX * fSize.fY };
 
 	std::vector<size_t>	safetyFactors	( iterations, 0 );
 
-	for ( int i = 0; i < iterations; i++ )
+	for ( int64_t i = 0; i < iterations; i++ )
 		safetyFactors[ i ]	= this->calculateSafetyFactorAfter( 1 );
 
 	return std::distance( safetyFactors.begin(), std::ranges::min_element( safetyFactors ) ) + 1;
@@ -59,7 +59,7 @@ size_t Space::findLowestSafetyFactorConfiguration()
 
 size_t Space::findSpaceWithFrameOfRobots()
 {
-	const int			iterations { fSize.fX * fSize.fY };
+	const int64_t		iterations { fSize.fX * fSize.fY };
 
 	std::vector<size_t>	safetyFactors	( iterations, 0 );
 
@@ -77,12 +77,12 @@ size_t Space::findSpaceWithFrameOfRobots()
 
 void Space::showAllGrids()
 {
-	const int	iterations	{ fSize.fX * fSize.fY };
+	const int64_t	iterations	{ fSize.fX * fSize.fY };
 
-	const int	startIter	{ 7573 };
+	const int64_t	startIter	{ 7573 };
 	this->calculateGridAfter( startIter );
 
-	for ( int i = startIter; i < iterations; i++ )
+	for ( int64_t i = startIter; i < iterations; i++ )
 	{
 		this->calculateGridAfter( 1 );
 		std::cout << "\n\nGrid after " << i << "\n";
@@ -92,9 +92,9 @@ void Space::showAllGrids()
 
 void Space::printGrid() const
 {
-	for ( int i = 0; i < fSize.fY; i++ )
+	for ( int64_t i = 0; i < fSize.fY; i++ )
 	{
-		for ( int j = 0; j < fSize.fX; j++ )
+		for ( int64_t j = 0; j < fSize.fX; j++ )
 			if ( const size_t& count { fRobotGrid[ i ][ j ] }; count == 0 )
 				std::cout << ' ';
 			else
@@ -106,8 +106,8 @@ void Space::printGrid() const
 
 void Space::calculateGridAfter(size_t seconds)
 {
-	for ( int i = 0; i < fSize.fY; i++ )
-		for ( int j = 0; j < fSize.fX; j++ )
+	for ( int64_t i = 0; i < fSize.fY; i++ )
+		for ( int64_t j = 0; j < fSize.fX; j++ )
 			fRobotGrid[ i ][ j ]	= 0;
 
 	for ( Robot& robot : fRobots )
